added TestCountVoteTransactions
